// api/src/auth/dto/confirmation-token.dto.ts
import { ApiProperty } from "@nestjs/swagger";
import { IsString } from "class-validator";

export class ConfirmationTokenDto {
  @ApiProperty({
    description:
      "The token used to verify your authorization event. This token is generated by the server and sent to the user's email.",
    example: "supersecre",
  })
  @IsString()
  token: string;
}